Fix failed tests, move js properties to frontend pom
